requirements.txt
Pillow
azure-cognitiveservices-vision-computervision
msrest